remove builded files from repo

+++ ComScroller/src/main/java/com/comscroller/ComScroller/api/EditorApiController.java
@@ -18,6 +18,7 @@ import com.comscroller.ComScroller.service.UserService;
 import com.comscroller.ComScroller.service.annotations.Role;
 import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,7 +35,6 @@ import org.springframework.web.bind.annotation.RestController;
  *
  * @author Catchy
  */
-
 @RestController
 @RequestMapping("/api/editor")
 public class EditorApiController {
@@ -51,54 +51,78 @@ public class EditorApiController {
     @Role({USER, ADMIN, MODERATOR})
     @GetMapping
     public ResponseEntity<Iterable<Games>> Games() {
-        Iterable<Games> games = editorService.list(userService.getUser());
-        return ResponseEntity.ok(games);
+        if (userService.isLoggedIn()) {
+            Iterable<Games> games = editorService.list(userService.getUser());
+            return ResponseEntity.ok(games);
+        }
+        return ResponseEntity.badRequest().build();
     }
-    
-    @Role({USER,ADMIN,MODERATOR})
+
+    @Role({USER, ADMIN, MODERATOR})
     @PutMapping("/finish")
     public ResponseEntity<Games> finish(@RequestBody Games game) {
-        return ResponseEntity.ok(editorService.finish(game));  
+        Users staff = userService.getUser();
+        if (userService.isAdmin(staff) || userService.isModerator(staff) || Objects.equals(staff.getId(), game.getOwnerid())) {
+
+            return ResponseEntity.ok(editorService.finish(game));
+        }
+        return ResponseEntity.badRequest().build();
     }
-    
-    @Role({USER,ADMIN,MODERATOR})
+
+    @Role({USER, ADMIN, MODERATOR})
     @PutMapping("/delete")
     public ResponseEntity<Boolean> delete(@RequestBody Games game, HttpServletResponse response, HttpServletRequest request) {
-        editorService.delete(game);
-        try {
-            response.sendRedirect("/api/editor/");
-        } catch (IOException e) {
+        Users staff = userService.getUser();
+        if (userService.isAdmin(staff) || userService.isModerator(staff) || Objects.equals(staff.getId(), game.getOwnerid())) {
+            editorService.delete(game);
+            try {
+                response.sendRedirect("/api/editor/");
+            } catch (IOException e) {
 
-            return ResponseEntity.badRequest().build();
+                return ResponseEntity.badRequest().build();
+            }
+            return ResponseEntity.ok(true);
         }
-        return ResponseEntity.ok(true);
+        return ResponseEntity.badRequest().build();
     }
-    
-    @Role({USER,ADMIN,MODERATOR})
+
+    @Role({USER, ADMIN, MODERATOR})
     @GetMapping("/create")
     public ResponseEntity<Games> create(HttpServletResponse response, HttpServletRequest request) {
-        return ResponseEntity.ok(new Games());
+        if (userService.isLoggedIn()) {
+            return ResponseEntity.ok(new Games());
+        }
+        return ResponseEntity.badRequest().build();
     }
-    
-    @Role({USER,ADMIN,MODERATOR})
+
+    @Role({USER, ADMIN, MODERATOR})
     @PostMapping("/create")
-    public ResponseEntity<Boolean> create(@RequestBody Games game, @RequestBody Iterable<Scenes> scenes,@RequestBody Iterable<Characters> characters, HttpServletResponse response, HttpServletRequest request) {
-        editorService.save(game, scenes, characters);       
-        return ResponseEntity.ok(true);
+    public ResponseEntity<Boolean> create(@RequestBody Games game, @RequestBody Iterable<Scenes> scenes, @RequestBody Iterable<Characters> characters, HttpServletResponse response, HttpServletRequest request) {
+        if (userService.isLoggedIn()) {
+            editorService.save(game, scenes, characters);
+            return ResponseEntity.ok(true);
+        }
+        return ResponseEntity.badRequest().build();
     }
-    
-    @Role({USER,ADMIN,MODERATOR})
+
+    @Role({USER, ADMIN, MODERATOR})
     @GetMapping("/edit/{id}")
-    public ResponseEntity<List<Object>> edit(@PathVariable int id,HttpServletResponse response, HttpServletRequest request) {
-        return ResponseEntity.ok(editorService.getAll(id));
+    public ResponseEntity<List<Object>> edit(@PathVariable int id, HttpServletResponse response, HttpServletRequest request) {
+        if (userService.isLoggedIn()) {
+            return ResponseEntity.ok(editorService.getAll(id));
+        }
+        return ResponseEntity.badRequest().build();
     }
-    
-    @Role({USER,ADMIN,MODERATOR})
+
+    @Role({USER, ADMIN, MODERATOR})
     @PostMapping("/edit/{id}")
-    public ResponseEntity<Boolean> edit(@RequestBody Games game, @RequestBody Iterable<Scenes> scenes,@RequestBody Iterable<Characters> characters, HttpServletResponse response, HttpServletRequest request) {
-        editorService.save(game, scenes, characters);
-       
-        return ResponseEntity.ok(true);
+    public ResponseEntity<Boolean> edit(@RequestBody Games game, @RequestBody Iterable<Scenes> scenes, @RequestBody Iterable<Characters> characters, HttpServletResponse response, HttpServletRequest request) {
+        Users staff = userService.getUser();
+        if (userService.isAdmin(staff) || userService.isModerator(staff) || Objects.equals(staff.getId(), game.getOwnerid())) {
+            editorService.save(game, scenes, characters);
+            return ResponseEntity.ok(true);
+        }
+        return ResponseEntity.badRequest().build();
     }
-    
+
 }
