+++ ComScroller/src/main/java/com/comscroller/ComScroller/api/UserApiController.java
@@ -26,7 +26,7 @@ public class UserApiController {
         this.userService = userService;
     }
 
-    @Role({USER,ADMIN,MODERATOR})
+    @Role({USER, ADMIN, MODERATOR})
     @GetMapping
     public ResponseEntity<Users> user() {
         if (userService.isLoggedIn()) {
@@ -34,25 +34,27 @@ public class UserApiController {
         }
         return ResponseEntity.badRequest().build();
     }
-    
-    @Role({ADMIN,MODERATOR})
+
+    @Role({ADMIN, MODERATOR})
     @GetMapping("/users")
     public ResponseEntity<Iterable<Users>> listAllUser() {
-        if (userService.isLoggedIn()) {
+        Users staff = userService.getUser();
+        if (userService.isAdmin(staff) || userService.isModerator(staff)) {
             Iterable<Users> users = userService.users();
             return ResponseEntity.ok(users);
         }
         return ResponseEntity.badRequest().build();
     }
-    
-    
-    @Role({ADMIN,MODERATOR})
+
+    @Role({ADMIN, MODERATOR})
     @GetMapping("/{username}")
     public ResponseEntity<Users> seeUser(@PathVariable String username) {
-        Users user = userService.getUser(username);
-        return ResponseEntity.ok(user);
+        Users staff = userService.getUser();
+        if (userService.isAdmin(staff) || userService.isModerator(staff)) {
+            return ResponseEntity.ok(userService.getUser(username));
+        }
+        return ResponseEntity.badRequest().build();
     }
-
 
     @PostMapping("/login")
     public ResponseEntity<Users> login(@RequestBody Users user) {
@@ -60,37 +62,53 @@ public class UserApiController {
             return ResponseEntity.ok(userService.login(user));
         } catch (UserNotValidException ev) {
             return ResponseEntity.badRequest().build();
-        }  catch (UserIsBannedException eb) {
+        } catch (UserIsBannedException eb) {
             return ResponseEntity.badRequest().build();
         }
-        
+
     }
-    
+
     @Role({ADMIN})
     @PutMapping("/role")
     public ResponseEntity<Users> changeRole(@RequestBody Users user, Users.Role role) {
-        return ResponseEntity.ok(userService.changeRole(user,role));  
+        Users staff = userService.getUser();
+        if (userService.isAdmin(staff)) {
+            return ResponseEntity.ok(userService.changeRole(user, role));
+        }
+        return ResponseEntity.badRequest().build();
     }
-    
-    @Role({ADMIN,MODERATOR})
+
+    @Role({ADMIN, MODERATOR})
     @PutMapping("/ban")
     public ResponseEntity<Users> ban(@RequestBody Users user) {
-        return ResponseEntity.ok(userService.ban(user));  
+        Users staff = userService.getUser();
+        if (userService.isAdmin(staff) || userService.isModerator(staff)) {
+            return ResponseEntity.ok(userService.ban(user));
+        }
+        return ResponseEntity.badRequest().build();
     }
-    
-    
+
     @Role({ADMIN})
     @PutMapping("/delete")
     public void delete(@RequestBody Users user) {
-        userService.delete(user);  
+        Users staff = userService.getUser();
+        if (userService.isAdmin(staff)) {
+            userService.delete(user);
+        }
     }
 
     @PostMapping("/registration")
     public ResponseEntity<Users> registration(@RequestBody Users user) {
-        try{
+        try {
             return ResponseEntity.ok(userService.registration(user));
-        } catch (SQLException | DataIntegrityViolationException e){
+        } catch (SQLException | DataIntegrityViolationException e) {
             return ResponseEntity.badRequest().build();
         }
+    }
+    
+    @GetMapping("/logout")
+    public ResponseEntity logout(@RequestBody Users user) {
+        this.userService.setUser(null);
+        return ResponseEntity.ok().build();
     }
 }
